no-venv → venv scenario
