+++ .travis.yml
@@ -14,6 +14,7 @@ script:
 env:
   - SCENARIO=newdep
   - SCENARIO=olddep
+  - SCENARIO=upgradetovenv
 
 notifications:
   webhooks: https://galaxy.ansible.com/api/v1/notifications/
