+++ molecule/resources/install-novenv.yml
@@ -1,0 +1,53 @@
+---
+- name: Install OMERO 5.4 without a virtualenv
+  hosts: all
+
+  roles:
+
+    - role: ome.postgresql
+      postgresql_install_server: true
+      postgresql_databases:
+        - name: omero
+      postgresql_users:
+        - user: omero
+          password: omero
+          databases: [omero]
+
+    - role: ansible-role-omero-server
+      omero_server_system_managedrepo_group: importer
+      omero_server_config_set:
+        Ice.IPv6: "0"
+      omero_server_virtualenv: false
+      # This will override host_vars (this is checked by the later tasks)
+      omero_server_release: "5.4"
+
+  tasks:
+
+    - name: get OMERO version
+      become: true
+      become_user: omero-server
+      command: /opt/omero/server/OMERO.server/bin/omero version
+      register: oldomero_version
+
+    - name: check that OMERO 5.4 was installed
+      assert:
+        that:
+          - oldomero_version.stdout.startswith('5.4.10-')
+
+    - name: check for presence of server venv
+      stat:
+        path: /opt/omero/server/venv
+      register: oldomero_venv_st
+
+    - name: check server venv does not exist
+      assert:
+        that:
+          - not oldomero_venv_st.stat.exists
+
+    - name: additional config file
+      copy:
+        content: >
+          config set omero.policy.binary_access -- "-read,-write,-image,-plate"
+        dest: /opt/omero/server/config/molecule-additional-config.omero
+      notify:
+        - restart omero-server
